app/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
streamlit==1.50.0
groq==1.0.0
psycopg2-binary==2.9.11
PyPDF2==3.0.1
sentence-transformers==5.1.2
numpy==2.0.2
python-dotenv==1.2.1
requests==2.32.5
